1.6.0 - fix #41
