+++ PrivateChannel.Front/Pages/Channel.razor
@@ -1,67 +1,47 @@
 ﻿@page "/channel/"
-@page "/channel/{ChannelId:guid}"
+@page "/channel/{CurrentChannelId:guid}"
 @inject PrivateChannelSvc.PrivateChannelSvcClient Client
 @inject NavigationManager NavigationManager
-@inject ISnackbar Snackbar
-@inject IJSRuntime JsInterop
 @inject IStringLocalizer<Channel> Localizer
+@inject IdentIconService IdentIcon
 
-<MudContainer Class="fill-height">
 
-    @if (ChannelId == null)
-    {
-        <div class="d-flex flex-column align-center mt-4">
-            <MudProgressCircular Color="Color.Primary" Size="Size.Large" Indeterminate="true" />
-            <MudText Typo="Typo.body1">@Localizer["InitializingNewChannelMessage"]</MudText>
-        </div>
-    }
-    else
-    {
-        <div class="d-flex fill-height flex-column gap-4 py-4">
 
-            @if (PeerPublicKey != null)
+
+
+<MudStack Row Class="fill-height" Spacing="0">
+    <MudPaper Class="d-none d-md-flex fill-height">
+        <MudNavMenu Class="overlow-x-hidden overflow-y-auto" Bordered Style="width:330px">
+            @foreach (Guid channel in Channels)
             {
-                <MudPaper Class="flex-grow-1 overflow-auto" Elevation="3">
-
-                    <MudStack Class="fill-height overflow-auto px-4 pb-4" Justify="Justify.FlexStart" AlignItems="AlignItems.Center" Row="false" Reverse Spacing="2">
-
-                        @foreach (var group in Messages.GroupBy(m => m.DateTime.Date).OrderByDescending(g => g.Key))
-                        {
-                            @foreach (MessageBase message in Messages.OrderByDescending(m => m.DateTime))
-                            {
-                                <MudPaper id="@(message.Id)" @key=message MaxWidth="90%" Class="@("pt-4 px-6 pb-2 " + (message.IsSender ? "mud-theme-secondary rounded-l-xl rounded-tr-xl align-self-end" : "mud-theme-primary rounded-r-xl rounded-tl-xl align-self-start"))">
-                                    @if (message is TextMessage textMessage)
-                                    {
-                                        <div class="d-flex flex-column gap-2" style="min-width:60px">
-                                            <pre style="white-space:pre-wrap"><MudText Typo="Typo.body1">@textMessage.Message</MudText></pre>
-                                            <MudText Class="@(" " + (message.IsSender ? "align-self-end" : "align-self-start"))" Typo="Typo.caption">@textMessage.DateTime.ToShortTimeString()</MudText>
-                                        </div>
-
-                                    }
-                                </MudPaper>
-                            }
-                            <MudText @key=group.Key Class="align-self-center" Typo="Typo.caption">@group.Key.ToLongDateString()</MudText>
-                        }
-                    </MudStack>
-
-                </MudPaper>
-
-                <MudPaper MinHeight="80px" Elevation="0">
-                    <MudTextField Margin="Margin.Dense" Variant="Variant.Filled" Placeholder="@Localizer["NewMessagePlaceholder"]" @bind-Value="TextToSend" Lines=@(Math.Max(2, Math.Min(TextToSend.Split('\n').Length ,6))) Adornment="Adornment.End" AdornmentIcon="@Icons.Material.Filled.Send" OnAdornmentClick="SendMessage" Immediate DisableUnderLine OnKeyDown="KeyDown"/>
-                </MudPaper>
-
+                <MudNavLink Icon="@IdentIcon.GetIconFromId(channel)" Match="NavLinkMatch.All" Href=@($"/channel/{channel}") Style="white-space:nowrap;overflow:hidden">@channel</MudNavLink>
             }
-            else
+            <MudButton FullWidth OnClick="AddChannel" Variant="Variant.Filled" Color="Color.Secondary" StartIcon="@Icons.Material.Filled.Add" Class="mt-2">@Localizer["AddChannelButton"]</MudButton>
+        </MudNavMenu>
+    </MudPaper>
+
+    <MudPaper Class="d-flex d-md-none fill-height">
+        <MudNavMenu Class="overlow-x-hidden overflow-y-auto" Bordered Style="width:60px">
+            @foreach (Guid channel in Channels)
             {
-                <div>
-                    <MudTextField ReadOnly Value="ChannelLink" Label="@Localizer["ShareLinkMessage"]" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.ContentCopy" OnAdornmentClick="CopyChannelLinkToClipboard" />
-                </div>
-                <div class="d-flex flex-column align-center mt-4">
-                    <MudProgressCircular Color="Color.Primary" Size="Size.Large" Indeterminate="true" />
-                    <MudText Typo="Typo.body1">@Localizer["WaitingPeerMessage"]</MudText>
-                </div>
+                <MudNavLink Icon="@IdentIcon.GetIconFromId(channel)" Match="NavLinkMatch.All" Href=@($"/channel/{channel}") Style="white-space:nowrap;overflow:hidden"></MudNavLink>
             }
+            <MudIconButton OnClick="AddChannel" Variant="Variant.Filled" Color="Color.Secondary" Icon="@Icons.Material.Filled.Add" Class="mt-2" Style="width:100%"></MudIconButton>
+        </MudNavMenu>
+    </MudPaper>
 
-        </div>
-    }
-</MudContainer>
+    <MudContainer Class="fill-height flex-grow-1">
+
+        @foreach (Guid channel in Channels)
+        {
+            <div class="@(channel == CurrentChannelId ? "d-block fill-height" : "d-none")">
+                <ChannelChat ChannelId="channel" />
+            </div>
+        }
+
+    </MudContainer>
+</MudStack>
+
+
+
+
