+++ PrivateChannel.Front/Pages/Channel.razor.cs
@@ -4,224 +4,40 @@ using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.JSInterop;
 using PrivateChannel.Front.Models;
+using System.Text;
 
 namespace PrivateChannel.Front.Pages;
 
 public partial class Channel
 {
-    #region Fields
-
-    private Guid? _CurrentChannel = null;
-
-    #endregion
-
     #region Properties
 
     [Parameter]
-    public Guid? ChannelId { get; set; }
+    public Guid? CurrentChannelId { get; set; }
 
-    private Guid? PeerId { get; set; }
-
-    private byte[]? PeerPublicKey { get; set; }
-    private List<byte>? PeerSessionKey { get; set; }
-
-    public string ChannelLink { get; set; } = String.Empty;
-
-    private List<MessageBase> Messages { get; set; } = new List<MessageBase>();
-
-    private SymetricKey? SymetricKey { get; set; }
-
-    private AsymetricKey? AsymetricKey { get; set; }
-
-    private byte[]? SessionKey { get; set; }
-
-    private bool SessionKeySended { get; set; }
-
-    private int SessionKeyRemainingUsage { get; set; }
-
-    private string TextToSend { get; set; } = string.Empty;
+    public List<Guid> Channels { get; set; } = new List<Guid>();
 
     #endregion
 
     #region Methods
 
-    protected override async Task OnParametersSetAsync()
+    protected override Task OnParametersSetAsync()
     {
-        if (ChannelId == null)
+        if (CurrentChannelId != null)
         {
-            try
+            if (Channels.Contains(CurrentChannelId.Value) == false)
             {
-                CreateChannelResponse reply = await Client.CreateChannelAsync(new CreateChannelRequest());
-                ChannelId = Guid.Parse(reply.ChannelId);
-                NavigationManager.NavigateTo($"/channel/{ChannelId}");
-
-            }
-            catch (Exception ex)
-            {
-                throw;
+                Channels.Add(CurrentChannelId.Value);
             }
         }
-        else if (_CurrentChannel != ChannelId)
-        {
-            _CurrentChannel = ChannelId;
-            ChannelLink = NavigationManager.Uri.ToString();
 
-            AsymetricKey = await JsInterop.InvokeAsync<AsymetricKey>("generateKeyPair");
-
-
-            _ = Task.Run(async () =>
-            {
-                try
-                {
-                    using var call = Client.ConnectToChannel(new ConnectToChannelRequest()
-                    {
-                        ChannelId = ChannelId.ToString(),
-                        PublicKey = ByteString.CopyFrom(AsymetricKey.PublicKey.ToArray())
-                    });
-
-                    await foreach (ConnectToChannelResponse connectToChannelResponse in call.ResponseStream.ReadAllAsync())
-                    {
-                        if (connectToChannelResponse.IsSelf)
-                        {
-                            PeerId = connectToChannelResponse.IsConnected ? Guid.Parse(connectToChannelResponse.PeerId) : null;
-
-                            _ = Task.Run(async () =>
-                            {
-                                try
-                                {
-                                    using var call = Client.GetMessages(new GetMessagesRequest()
-                                    {
-                                        ChannelId = ChannelId.ToString(),
-                                        PeerId = PeerId.ToString()
-                                    });
-
-                                    await foreach (GetMessagesResponse getMessageResponse in call.ResponseStream.ReadAllAsync())
-                                    {
-                                        if (PeerSessionKey == null || getMessageResponse.SessionKey.ToByteArray().Length > 0)
-                                        {
-                                            PeerSessionKey = getMessageResponse.SessionKey.ToByteArray().ToList();
-                                        }
-
-                                        string message = await JsInterop.InvokeAsync<string>("decryptWithPrivateKeyAndSymmetricKey",
-                                            getMessageResponse.EncryptedMessage.ToByteArray().ToList()
-                                            , getMessageResponse.IV.ToByteArray().ToList()
-                                            , PeerSessionKey
-                                            , AsymetricKey.PrivateKey);
-
-
-                                        int scrollPosition = await JsInterop.InvokeAsync<int>("scrollPosition", "channel-pannel");
-
-                                        MessageBase msg = new TextMessage()
-                                        {
-                                            Id = Guid.NewGuid(),
-                                            Message = message
-                                        };
-
-                                        Messages.Add(msg);
-                                        await InvokeAsync(StateHasChanged);
-
-                                        if (scrollPosition >= 0)
-                                        {
-                                            await JsInterop.InvokeVoidAsync("scrollIntoView", msg.Id);
-                                        }
-                                    }
-                                }
-                                catch (Exception ex)
-                                {
-
-                                    throw;
-                                }
-                            });
-                        }
-                        else
-                        {
-                            if (connectToChannelResponse.IsConnected)
-                            {
-                                PeerPublicKey = connectToChannelResponse.PublicKey.ToByteArray();
-                                await GenerateSessionKey();
-                            }
-                            else
-                            {
-                                PeerPublicKey = null;
-                                SymetricKey = null;
-                                SessionKey = null;
-                                SessionKeySended = false;
-                            }
-                        }
-
-                        await InvokeAsync(StateHasChanged);
-                    }
-                }
-                catch (Exception)
-                {
-
-                    throw;
-                }
-            });
-        }
+        return base.OnParametersSetAsync();
     }
 
-    private async Task GenerateSessionKey()
+    private async Task AddChannel()
     {
-        SymetricKey = await JsInterop.InvokeAsync<SymetricKey>("generateAndEncryptSymmetricKey", PeerPublicKey.ToList());
-        SessionKey = SymetricKey.Encrypted.ToArray();
-        SessionKeyRemainingUsage = 20;
-        SessionKeySended = false;
-    }
-
-    private async Task CopyChannelLinkToClipboard()
-    {
-        await JsInterop.InvokeVoidAsync("navigator.clipboard.writeText", ChannelLink);
-        Snackbar.Add(Localizer["SnackbarLinkCopied"]);
-    }
-
-    private async Task SendMessage()
-    {
-        if (SymetricKey == null || SessionKeyRemainingUsage <= 0)
-        {
-            await GenerateSessionKey();
-        }
-
-        if (SymetricKey != null && SessionKeyRemainingUsage > 0)
-        {
-            string messageToSend = TextToSend;
-
-            EncryptedMessage encryptedMessage = await JsInterop.InvokeAsync<EncryptedMessage>("encryptWithSymmetricKey", messageToSend, SymetricKey.Raw);
-
-            await Client. SendMessageAsync(new SendMessageRequest()
-            {
-                ChannelId = ChannelId.ToString(),
-                PeerId = PeerId.ToString(),
-                SessionKey = ByteString.CopyFrom(SessionKeySended ? new byte[0] : SessionKey),
-                EncryptedMessage = ByteString.CopyFrom(encryptedMessage.EncryptedData.ToArray()),
-                IV = ByteString.CopyFrom(encryptedMessage.IV.ToArray())
-            });
-
-            SessionKeySended = true;
-
-            MessageBase msg = new TextMessage()
-            {
-                Id = Guid.NewGuid(),
-                Message = messageToSend,
-                IsSender = true
-            };
-
-            Messages.Add(msg);
-            TextToSend = string.Empty;
-
-            await InvokeAsync(StateHasChanged);
-
-            SessionKeyRemainingUsage--;
-            await JsInterop.InvokeVoidAsync("scrollIntoView", msg.Id);
-        }
-    }
-
-    private async Task KeyDown(KeyboardEventArgs args)
-    {
-        if (args.ShiftKey == false && args.Key == "Enter")
-        {
-            await SendMessage();
-        }
+        CreateChannelResponse reply = await Client.CreateChannelAsync(new CreateChannelRequest());
+        NavigationManager.NavigateTo($"/channel/{Guid.Parse(reply.ChannelId)}");
     }
 
     #endregion
