+++ PrivateChannel.Front/_Imports.razor
@@ -11,4 +11,5 @@
 @using PrivateChannel.Front.Layouts
 @using PrivateChannel.Front.Models
 @using PrivateChannel.Front.Components
+@using PrivateChannel.Front.Services
 @using MudBlazor
